Added typedoc JSON provider
