+++ package.json
@@ -22,7 +22,8 @@
   "peerDependencies": {},
   "dependencies": {
     "fs-extra": "7.0.1",
-    "glob": "7.1.3"
+    "glob": "7.1.3",
+    "typedoc": "0.14.2"
   },
   "devDependencies": {
     "eslint": "4.19.1",
